requirements.txt
 streamlit as st
 requests
 pandas as pd
 matplotlib.pyplot as plt